Backend - Extract RequisitionItem into new service
