+++ backend/src/main/java/com/project/requisition/service/RequisitionService.java
@@ -2,11 +2,8 @@ package com.project.requisition.service;
 
 import com.project.requisition.constants.RequisitionStatus;
 import com.project.requisition.entity.RequisitionEntity;
-import com.project.requisition.entity.RequisitionItemEntity;
 import com.project.requisition.exception.NotFoundException;
-import com.project.requisition.model.request.RequisitionItemRequest;
 import com.project.requisition.model.request.RequisitionRequest;
-import com.project.requisition.repository.RequisitionItemRepository;
 import com.project.requisition.repository.RequisitionRepository;
 import com.project.requisition.utils.CommonUtils;
 import com.project.requisition.utils.ValidateUtils;
@@ -17,7 +14,6 @@ import org.springframework.stereotype.Service;
 
 import java.time.Clock;
 import java.time.ZonedDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -28,8 +24,9 @@ import static com.project.requisition.constants.ErrorConstant.REQUISITION_NOT_FO
 @RequiredArgsConstructor
 public class RequisitionService {
 
+    private final RequisitionItemService requisitionItemService;
+
     private final RequisitionRepository requisitionRepository;
-    private final RequisitionItemRepository requisitionItemRepository;
     private final Clock systemClock;
 
     public List<RequisitionEntity> getAllRequisitions() {
@@ -66,7 +63,7 @@ public class RequisitionService {
         requisitionEntity.setLastUpdatedTimestamp(currentDateTime);
 
         requisitionRepository.save(requisitionEntity);
-        createRequisitionItem(requisitionId, operatorId, requisitionRequest.itemList(), currentDateTime);
+        requisitionItemService.createRequisitionItem(requisitionId, operatorId, requisitionRequest.itemList(), currentDateTime);
 
         return requisitionEntity;
     }
@@ -101,7 +98,7 @@ public class RequisitionService {
 
         requisitionRepository.save(requisitionEntity);
         if (!CommonUtils.isListNullOrEmpty(requisitionRequest.itemList())) {
-            patchRequisitionItem(requisitionEntity.getRequisitionId(), requisitionEntity.getLastOperatorId(), requisitionRequest.itemList(), currentDateTime);
+            requisitionItemService.patchRequisitionItem(requisitionEntity.getRequisitionId(), requisitionEntity.getLastOperatorId(), requisitionRequest.itemList(), currentDateTime);
         }
 
         return requisitionEntity;
@@ -117,64 +114,7 @@ public class RequisitionService {
                     return new NotFoundException(err);
                 });
         requisitionRepository.delete(requisitionEntity);
-        requisitionItemRepository.deleteByRequisitionId(requisitionId);
-    }
-
-    private void createRequisitionItem(UUID requisitionId, UUID operatorId, List<RequisitionItemRequest> itemList, ZonedDateTime currentDateTime) {
-        var requisitionItemEntityList = new ArrayList<RequisitionItemEntity>();
-        itemList.forEach(item -> {
-            var requisitionItemEntity = populateRequisitionItemEntity(requisitionId, item, operatorId, currentDateTime);
-            requisitionItemEntityList.add(requisitionItemEntity);
-        });
-        requisitionItemRepository.saveAll(requisitionItemEntityList);
-    }
-
-    @Transactional
-    public void patchRequisitionItem(UUID requisitionId, UUID operatorId, List<RequisitionItemRequest> updatedList, ZonedDateTime currentDateTime) {
-        var existingList = requisitionItemRepository.findByRequisitionId(requisitionId);
-        var deleteList = new ArrayList<UUID>();
-
-        // Process update existing item when edit requisition
-        for (var updated : updatedList) {
-            for (var existing : existingList) {
-                if (existing.getItemId().equals(UUID.fromString(updated.itemId()))) {
-                    if (updated.quantity() == 0) {
-                        // Process delete item when edit requisition
-                        deleteList.add(existing.getItemId());
-                    }
-                    existing.setQuantity(updated.quantity());
-                    existing.setLastOperatorId(operatorId);
-                    existing.setLastUpdatedTimestamp(currentDateTime);
-                }
-            }
-        }
-
-        // Process add new item when edit requisition
-        if (existingList.size() != updatedList.size()) {
-            updatedList.stream()
-                    .filter(updated -> existingList.stream().noneMatch(existing -> existing.getItemId().equals(UUID.fromString(updated.itemId()))))
-                    .forEach(item -> {
-                        var requisitionItemEntity = populateRequisitionItemEntity(requisitionId, item, operatorId, currentDateTime);
-                        existingList.add(requisitionItemEntity);
-                    });
-        }
-
-        requisitionItemRepository.saveAll(existingList);
-        requisitionItemRepository.deleteByRequisitionId(requisitionId, deleteList);
-    }
-
-    private RequisitionItemEntity populateRequisitionItemEntity(
-            UUID requisitionId, RequisitionItemRequest requisitionItemRequest, UUID operatorId, ZonedDateTime currentDateTime
-    ) {
-        var requisitionItemEntity = new RequisitionItemEntity();
-        requisitionItemEntity.setRequisitionItemId(UUID.randomUUID());
-        requisitionItemEntity.setRequisitionId(requisitionId);
-        requisitionItemEntity.setItemId(UUID.fromString(requisitionItemRequest.itemId()));
-        requisitionItemEntity.setQuantity(requisitionItemRequest.quantity());
-        requisitionItemEntity.setLastOperatorId(operatorId);
-        requisitionItemEntity.setCreatedTimestamp(currentDateTime);
-        requisitionItemEntity.setLastUpdatedTimestamp(currentDateTime);
-        return requisitionItemEntity;
+        requisitionItemService.deleteRequisitionItemByRequisitionId(requisitionId);
     }
 
 }
